Added security enforcements -verify some actions can be done only by admin or by specific user

+++ src/main/java/controllers/ApiDeviceType.java
@@ -33,7 +33,7 @@ public class ApiDeviceType{
 		Gson gson = new Gson();
 		
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			deviceType = gson.fromJson(deviceTypeJson, DeviceType.class);
 			DeviceTypeHander.insertNewDeviceType(deviceType.getName(), deviceType.getPicData());
 			response = Response.ok(GenericResponse.ok(DeviceTypeHander.DEVICES_TYPE_CREATE_SUCCESS_MESSAGE)).build();
@@ -54,7 +54,7 @@ public class ApiDeviceType{
 		Gson gson = new Gson();
 		
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			deviceType = gson.fromJson(deviceTypeJson, DeviceType.class);
 			DeviceTypeHander.updateDeviceType(deviceType);
 			response = Response.ok(GenericResponse.ok(DeviceTypeHander.DEVICES_TYPE_UPDATE_SUCCESS_MESSAGE)).build();
@@ -72,7 +72,7 @@ public class ApiDeviceType{
 	public Response delete(@Context HttpServletRequest req,@PathParam("deviceTypeID") int deviceTypeID){
 		Response response = null;
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			DeviceTypeHander.deleteDeviceType(deviceTypeID);
 			response = Response.ok(GenericResponse.ok(DeviceTypeHander.DEVICES_TYPE_DELETE_SUCCESS_MESSAGE)).build();
 		}
@@ -90,7 +90,7 @@ public class ApiDeviceType{
 		Response response = null;
 		DeviceType deviceType = null;
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			deviceType = DeviceTypeHander.getDeviceTypeByID(deviceTypeID);
 			response = Response.ok(GenericResponse.ok(deviceType)).build();
 		}
@@ -108,7 +108,7 @@ public class ApiDeviceType{
 		Response response = null;
 		List<DeviceType> devicesTypes = null;
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			devicesTypes = DeviceTypeHander.getAllDevicesTypes();
 			response = Response.ok(GenericResponse.ok(devicesTypes)).build();
 		}
