+++ src/main/java/controllers/ApiUser.java
@@ -30,7 +30,7 @@ import modelObjects.User.UserType;
 
 @Path("/user")
 public class ApiUser{
-	
+
 	@GET
 	@Path("/authenticateuser")
 	@Produces(MediaType.APPLICATION_JSON)
@@ -39,44 +39,30 @@ public class ApiUser{
 		Response response = null;
 
 		try {
-			if(SessionHandler.isAuthUser(req) == false) {
-				throw new Exception("Access Denied - Login First");
-			}
-
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			UserType type = SessionHandler.getType(req);
-
 			user.setUserID(SessionHandler.getId(req));
 			user.setType(type);
-
 			String fullName = SessionHandler.getFullname(req);
 			String[] fullNameArr = fullName.split(" ");
-
 			user.setFirstname(fullNameArr[0]);
 			user.setLastname(fullNameArr[1]);
-
 			response = Response.ok(GenericResponse.ok(user)).build();
-
 		} catch(Exception e) {
 			response = Response.ok(GenericResponse.error(e.getMessage())).build();
 		}
 
 		return response;
 	}	
-	
+
 	@POST
 	@Path("/register")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response insert(@Context javax.servlet.http.HttpServletRequest req  , @FormParam("username") String username,@FormParam("password") String password, @FormParam("firstname") String firstname, @FormParam("lastname") String lastname,@FormParam("email") String email,@FormParam("mobile") String mobile) {
 		// String hashedPassword = PasswordHash.hash(password);
-
 		Response response = null;
 		try {
-//			if(SessionHandler.isAuthUser(req) == false){
-//				throw new Exception("Access Denied - Login First");
-//			}
-//			if(SessionHandler.getType(req) != UserType.Admin){
-//				throw new Exception("Access Denied - Only an admin can create user");
-//			}
+			SessionHandler.verifyAdminRequest(req);
 			UserHandler.insertNewUser(username.toLowerCase() , password, firstname , lastname,email,mobile);   
 			response = Response.ok(GenericResponse.ok("user has been created successfully")).build();
 		} 
@@ -86,35 +72,6 @@ public class ApiUser{
 		return response;
 	}
 
-//	@POST
-//	@Path("/login")
-//	@Produces(MediaType.APPLICATION_JSON)
-//	public Response login(@Context javax.servlet.http.HttpServletRequest req  , @FormParam("username") String username,@FormParam("password") String password) {
-//		Response response = null;
-//		try{
-//			if(SessionHandler.isAuthUser(req) == true ) {
-//				throw new Exception("Access Denied - You Are Already Logged In");
-//			}
-//			User user = UserHandler.getUserByUsername(username);
-//			if(user==null){
-//				throw new Exception("Wrong UserName Or Password");
-//			}
-//			if(user.getPassword().equals(password)){
-//				user.setPassword("********");
-//				
-//				SessionHandler.authUser(user.getUserID(),user.getType(),user.getFirstname(),user.getLastname(), req);	
-//				response = Response.ok(GenericResponse.ok(user)).build();
-//			}
-//			else{
-//				response = Response.ok(GenericResponse.error(Response.Status.UNAUTHORIZED)).build();
-//			}
-//		}
-//		catch(Exception ex){
-//			response = Response.ok(GenericResponse.error(ex.getMessage())).build();
-//		}
-//
-//		return response;
-//	}
 	@POST
 	@Path("/login")
 	@Produces(MediaType.APPLICATION_JSON)
@@ -133,7 +90,7 @@ public class ApiUser{
 			}
 			if(user.getPassword().equals(password)){
 				user.setPassword("********");
-				
+
 				SessionHandler.authUser(user.getUserID(),user.getType(),user.getFirstname(),user.getLastname(), req);	
 				response = Response.ok(GenericResponse.ok(user)).build();
 			}
@@ -147,24 +104,22 @@ public class ApiUser{
 
 		return response;
 	}
-	
+
 	@POST
 	@Path("/logout")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response logout(@Context HttpServletRequest req){
 		Response response = null;
-		
+
 		try{
-			if (SessionHandler.isAuthUser(req) == false){
-				throw new Exception("Access Denied - Login First");
-			}
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			SessionHandler.unAuthUser(req);
 			response = Response.ok(GenericResponse.ok(UserHandler.USER_LOGOUT_SUCCESS_MESSAGE)).build();
 		}
 		catch(Exception ex){
 			response = Response.ok(GenericResponse.error(ex.getMessage())).build();
 		}
-		
+
 		return response;
 	}
 
@@ -177,9 +132,10 @@ public class ApiUser{
 		Gson gson = new Gson();
 
 		try{
-//			SessionHandler.verifyUserIsAuthenticated(req);
+						
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			user = gson.fromJson(userJson, User.class);
-			//Check if admin or request to update his user
+			SessionHandler.verifyUserIsAuthorized(req, user.getUserID());
 			UserHandler.updateUser(user);
 			response = Response.ok(GenericResponse.ok(UserHandler.USER_UPDATE_SUCCESS_MESSAGE)).build();
 		}
@@ -192,7 +148,7 @@ public class ApiUser{
 
 		return response;
 	}
-	
+
 	@PUT
 	@Path("/change_password")
 	@Produces(MediaType.APPLICATION_JSON)
@@ -203,13 +159,14 @@ public class ApiUser{
 		// In case admin wants to change password, he doesn't know that is the old password 
 		//
 		//===============================
-		
+
 		//json structure :{userID:X ,oldPassword: X ,newPassword: X}
 		//example: {userID:4,oldPassword:"12345678",newPassword:"1234"}
 		try{
-//			SessionHandler.verifyUserIsAuthenticated(req);
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			JSONObject jsonObject = new JSONObject(passwordJson);
 			int userID = jsonObject.getInt("userID");
+			SessionHandler.verifyUserIsAuthorized(req, userID);
 			String oldPassword = jsonObject.getString("oldPassword");
 			String newPassword = jsonObject.getString("newPassword");
 			UserHandler.changeUserPassword(userID, oldPassword, newPassword);
@@ -224,16 +181,15 @@ public class ApiUser{
 
 		return response;
 	}
-	
+
 	@PUT
 	@Path("/reset_password")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response resetPassword(@Context HttpServletRequest req, String passwordJson){
 		Response response = null;
 
-		//example: {userID:4}
 		try{
-//			SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			JSONObject jsonObject = new JSONObject(passwordJson);
 			int userID = jsonObject.getInt("userID");
 			UserHandler.resetUserPassword(userID, UserHandler.USER_DEFAULT_RESET_PASSWORD);
@@ -257,19 +213,15 @@ public class ApiUser{
 		User user = null;
 		Response response = null;
 		try{
-			//SessionHandler.verifyUserIsAuthenticated(req);
+			SessionHandler.verifyAuthenticatedUserRequest(req);
 			user = UserHandler.getUserByUsername(username);
 			user.setPassword("********");
+			response = Response.ok(GenericResponse.ok(user)).build();
 		}
 		catch (Exception e) {
 			response = Response.ok(GenericResponse.error(e.getMessage())).build();
 		}
-		if (user == null ){
-			response = Response.ok(GenericResponse.error("user wasn't found")).build();
-		}
-		else{
-			response = Response.ok(GenericResponse.ok(user)).build();
-		}
+
 		return response;
 	}
 	@GET
@@ -280,7 +232,7 @@ public class ApiUser{
 		List<User> users = null;
 		Response response = null;
 		try{
-			//SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			users = UserHandler.getAllUsers();
 			response = Response.ok(GenericResponse.ok(users)).build();
 		}
@@ -297,7 +249,7 @@ public class ApiUser{
 	public Response delete(@Context HttpServletRequest req,@PathParam("userID") int userID){
 		Response response = null;
 		try{
-			//SessionHandler.verifyAdminRequest(req);
+			SessionHandler.verifyAdminRequest(req);
 			UserHandler.deleteUser(userID);
 			response = Response.ok(GenericResponse.ok(UserHandler.USER_UPDATE_DELETE_MESSAGE)).build();
 		}
@@ -307,7 +259,4 @@ public class ApiUser{
 
 		return response;
 	}
-	
-	
-
 }
